requirements.txt
argcomplete==1.8.2
asn1crypto==0.22.0
base58==0.2.4
boto3==1.4.4
botocore==1.5.40
cffi==1.10.0
click==6.7
cryptography==1.8.1
Django==1.11.1
django-debug-toolbar==1.8
django-nose==1.4.4
docutils==0.13.1
durationpy==0.4
future==0.16.0
futures==3.1.1
hjson==2.0.2
idna==2.5
jmespath==0.9.2
kappa==0.6.0
lambda-packages==0.15.0
nose==1.3.7
packaging==16.8
placebo==0.8.1
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.5.0
pyOpenSSL==17.0.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.0
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
requests==2.14.2
s3transfer==0.1.10
six==1.10.0
sqlparse==0.2.3
toml==0.9.2
tqdm==4.11.2
troposphere==1.9.3
twilio==6.1.2
Unidecode==0.4.20
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.42.0
